In this project, an Agentic AI Application has been built for YouTube Video Summary Extraction in Python with Phidata, YouTubeTools and Llama4 LLM (with Groq cloud)
In this project, an Agentic AI Application has been built for YouTube Video Summary Extraction in Python with Phidata, YouTubeTools and Llama4 LLM (with Groq cloud). For building the app, Streamlit library is used in python. YouTubeTools enables an AI Agent to access captions and metadata of YouTube videos when provided with a video URL link. In the App, user needs to provide only the YouTube Video Link, then they just need to click on the "Extract Video Summary" button. Our AI Agent, with the help of YouTube Tools will extract the detailed video summary and display it there.